fix(task-2): update type sizes
uint8_t seemed too fragile to use and its just annoying, I won't
use this just to show that I know sizes of types, this is stupid

+++ task-2/include/simulation/types.hpp
@@ -6,7 +6,7 @@
 namespace sim {
 
 /// Index of the product type
-using product_type_t = uint8_t;
+using product_type_t = uint16_t;
 
 /// Index of the product
 using product_index_t = uint32_t;
@@ -23,10 +23,10 @@ struct Product {
 using product_t = Product;
 
 /// Index of the operation type
-using operation_t = uint8_t;
+using operation_t = uint16_t;
 
 /// Index of the machine
-using machine_t = uint8_t;
+using machine_t = uint16_t;
 
 /// Operation time
 using optime_t = uint16_t;
